Split mailbox module by concern
