+++ crates/shelterwood/src/mailbox/errors.rs
@@ -1,0 +1,152 @@
+use std::fmt;
+
+use crate::{ChildId, Incarnation};
+
+/// The kind of a failed actor send.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[non_exhaustive]
+pub enum SendErrorKind {
+    /// The membership has no accepting incarnation right now.
+    NotRunning,
+    /// A bounded FIFO mailbox has no free capacity.
+    Full,
+    /// The membership is terminal.
+    Terminated,
+    /// A timed send was withdrawn before acceptance.
+    TimedOut,
+}
+
+/// A failed send with its recoverable message and identity evidence.
+pub struct SendError<M> {
+    /// Target actor id.
+    pub actor_id: ChildId,
+    /// Incarnation required by the error-kind identity table.
+    pub incarnation_observed: Option<Incarnation>,
+    /// Message proven not to have been accepted.
+    pub message: M,
+    /// Failure category.
+    pub kind: SendErrorKind,
+}
+
+impl<M> fmt::Debug for SendError<M> {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter
+            .debug_struct("SendError")
+            .field("actor_id", &self.actor_id)
+            .field("incarnation_observed", &self.incarnation_observed)
+            .field("message", &"<recoverable message>")
+            .field("kind", &self.kind)
+            .finish()
+    }
+}
+
+impl<M> fmt::Display for SendError<M> {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(formatter, "send to {} failed: {}", self.actor_id, self.kind)
+    }
+}
+
+impl<M> std::error::Error for SendError<M> {}
+
+impl fmt::Display for SendErrorKind {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter.write_str(match self {
+            Self::NotRunning => "actor is not accepting messages",
+            Self::Full => "mailbox is full",
+            Self::Terminated => "actor membership is terminal",
+            Self::TimedOut => "acceptance deadline elapsed",
+        })
+    }
+}
+
+/// The kind of a failed request/reply call.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[non_exhaustive]
+pub enum CallErrorKind {
+    /// The membership terminalized before acceptance.
+    Terminated,
+    /// The request was withdrawn before acceptance.
+    AcceptanceTimedOut,
+    /// The request was accepted but its response missed the deadline.
+    ResponseTimedOut,
+    /// The accepted request's reply capability was dropped unanswered.
+    ReplyDropped,
+}
+
+/// A failed request/reply call with retry-discipline identity evidence.
+///
+/// [`CallErrorKind::AcceptanceTimedOut`] proves that the request was not
+/// accepted and is safe to retry. [`CallErrorKind::ResponseTimedOut`] means
+/// that it was accepted with an unknown outcome and must be reconciled rather
+/// than blindly retried. A [`CallErrorKind::ReplyDropped`] retry is safe only
+/// for an idempotent operation, under one overall deadline, after snapshots or
+/// lifecycle events show an incarnation that
+/// [`supersedes`](Incarnation::supersedes) `incarnation_observed`.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct CallError {
+    /// Target actor id.
+    pub actor_id: ChildId,
+    /// Accepting or observed incarnation, as determined by [`CallErrorKind`].
+    pub incarnation_observed: Option<Incarnation>,
+    /// Failure category.
+    pub kind: CallErrorKind,
+}
+
+impl fmt::Display for CallError {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(formatter, "call to {} failed: {}", self.actor_id, self.kind)
+    }
+}
+
+impl std::error::Error for CallError {}
+
+impl fmt::Display for CallErrorKind {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter.write_str(match self {
+            Self::Terminated => "actor membership terminalized before acceptance",
+            Self::AcceptanceTimedOut => "request acceptance deadline elapsed",
+            Self::ResponseTimedOut => "response deadline elapsed after acceptance",
+            Self::ReplyDropped => "reply capability was dropped unanswered",
+        })
+    }
+}
+
+/// A successful call result and the incarnation that accepted its request.
+#[derive(Clone, Eq, PartialEq)]
+pub struct Replied<T> {
+    /// Reply value.
+    pub value: T,
+    /// Incarnation that accepted the request.
+    pub incarnation: Incarnation,
+}
+
+impl<T> fmt::Debug for Replied<T> {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter
+            .debug_struct("Replied")
+            .field("value", &"<reply value>")
+            .field("incarnation", &self.incarnation)
+            .finish()
+    }
+}
+
+/// Failure of a standalone reply receiver.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[non_exhaustive]
+pub enum ReplyError {
+    /// The reply capability was dropped unanswered.
+    Dropped,
+    /// The response deadline elapsed.
+    Timeout,
+}
+
+impl fmt::Display for ReplyError {
+    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
+        formatter.write_str(match self {
+            Self::Dropped => "reply capability was dropped unanswered",
+            Self::Timeout => "reply deadline elapsed",
+        })
+    }
+}
+
+impl std::error::Error for ReplyError {}
